Added check mode support
Also changed the behavior of api_query() by allowing it to return None
for request that resuts in status: False
This change will allow the module to handle event where the defined
server does not exist.

Also modified get_server_by_name() to allow it to handle case where
defined server does not exist.

Furthermore modified galaxy.yml to allow it to be uploaded to ansible
galaxy. Currently the repository link is not real, added my name for
author information, changed version to 0.0.1 to reflect work in progress,
and added a GPL liecense_file called COPYING.